World data and texture update
Updated ARENA01.RMB with street festival and building data.
Updated and finished layout for the gladiator school
Created layout for the tavern
Added NPC textures for the flamebreather and two sitting NPCs.

Buildings have a description after the building number to make it easier to import their data into the arena RMB.